Remove the repack tests
The configure_and_write_nwbfile tests cover the fix.
